Configure hugo pipes for postcss and javascript, disables gulp js and css, moved css and js dirs into assets
