+++ src/assets/css/_fills.css
@@ -1,0 +1,1 @@
+.fill-current-color { fill: currentcolor };
