+++ src/assets/css/_footer.css
@@ -1,0 +1,7 @@
+.footer-component li > .link {
+  padding-bottom: .1rem;
+}
+
+.footer-component li > .link:hover {
+  border-bottom: 1px solid currentcolor;
+}
