+++ src/assets/css/_forms.css
@@ -1,0 +1,19 @@
+input {
+  height: 44px;
+  background: $black;
+  border-color: $black;
+  padding: 0 $spacing-small;
+  outline: none;
+  border-style:solid;
+}
+
+form.form-white input {
+  background: #222323;
+  border-color: #222323;
+  outline: none;
+}
+
+form.form-white input:hover {
+  color: #fff;
+  border-color: #000;
+}
