+++ src/assets/css/_manifesto.css
@@ -1,0 +1,22 @@
+/* Play fair. */
+#manifesto-header h1 > strong {
+  display: block;
+  font-weight: normal;
+  color: var(--black);
+  font-size: 30vw;
+  line-height: 1;
+  letter-spacing: -.05em;
+  margin-left: -.066em;
+}
+
+@media (min-width: 30rem) {
+  #manifesto-header h1 > strong {
+    font-size: 17.666vw;
+  }
+}
+
+@media (min-width: 60rem) {
+  #manifesto-header h1 > strong {
+    font-size: 10.395625rem; /* 166.333px */
+  }
+}
