+++ src/assets/css/_marketing.css
@@ -1,0 +1,116 @@
+article:focus {
+  outline: 0;
+}
+
+#sticky-sidebar {
+  position: sticky;
+  top: calc(50% - 192px);
+}
+
+// paragraphs
+.paragraphs {
+  margin-bottom: 3.667rem;
+  padding: 0;
+}
+
+.paragraphs:last-child {
+  margin-bottom: 0;
+}
+
+@media screen and (min-width: 60em) {
+  .paragraphs {
+    margin-bottom: 5.867rem;
+  }
+}
+
+// pricing
+
+.pricing-dots {
+  box-sizing: content-box;
+  width: 100%;
+  padding-left: 40%;
+  margin-right: 0;
+}
+
+.pricing-paragraph {
+  color: $black;
+  position: relative;
+  margin-left: -40%;
+}
+
+@media screen and (min-width: 60em) {
+  .pricing-dots {
+    width: 1026px;
+    padding-left: 0;
+    margin-right: -346px;
+  }
+
+  .pricing-paragraph {
+    color: $white;
+    position: absolute;
+    top:40%;
+    right: 346px;
+    margin-left: 0;
+  }
+}
+
+.pricing-lines {
+  height: 1px;
+  margin-right: -490px;
+  background-image: linear-gradient(to right, black 33%, rgba(255,255,255,0) 0%);
+  background-position: bottom;
+  background-size: 4px 1px;
+  background-repeat: repeat-x;
+}
+
+// TGE Banner
+.f-tge {
+  font-size: 1.44rem;
+  line-height: 1.73rem;
+  letter-spacing: -1px;
+  padding: 0;
+}
+
+@media screen and (min-width: 60em) {
+  .f-tge {
+    font-size: 2.4rem;
+    line-height: 2.9rem;
+    letter-spacing: -2px;
+  }
+}
+
+.f-tge--green {
+  color: #6CF289;
+  font-weight: 600;
+  padding:0;
+}
+
+// Logo images
+
+.logoimage-wrapper {
+  padding-bottom: 48%
+}
+
+@media screen and (min-width: 30em) {
+  .logoimage-wrapper {
+    padding-bottom: 19%
+  }
+}
+
+.logoimage-wrapper img.logoimage {
+  min-width: 200px;
+  min-height: 200px;
+  left: 50%;
+  top: 50%;
+  margin-top: -100px;
+  margin-left: -100px;
+}
+
+@media screen and (min-width: 30em) and (max-width: 60em) {
+  .logoimage-wrapper img.logoimage {
+    min-width: 160px;
+    min-height: 160px;
+    margin-top: -80px;
+    margin-left: -80px;
+  }
+}
