+++ src/assets/css/_social-icons.css
@@ -1,0 +1,44 @@
+.facebook, .twitter, .instagram, .youtube, .github, .gitlab, .linkedin, .medium {
+  fill: #BABABA;
+}
+
+.new-window {
+  opacity: 0;
+  display: inline-block;
+  vertical-align: top;
+}
+.link-transition:hover .new-window{
+  opacity: 1;
+}
+
+.facebook:hover {
+  fill: #3b5998;
+}
+
+.twitter:hover {
+  fill: #1da1f2;
+}
+
+.instagram:hover {
+  fill: #e1306c;
+}
+
+.youtube:hover {
+  fill: #cd201f;
+}
+
+.github:hover {
+  fill: #6cc644;
+}
+
+.gitlab:hover {
+  fill: #FC6D26;
+}
+
+.linkedin:hover {
+  fill: #0077b5
+}
+
+.medium:hover {
+  fill: #0077b5
+}
