+++ src/assets/css/_syntax.css
@@ -1,0 +1,73 @@
+.chroma .hll { background-color: #ffffcc }
+.chroma  { background: #282828; color: #ebdbb2; background-color: #282828 }
+.chroma .c { color: #928374; font-style: italic; background-color: #282828 } /* Comment */
+.chroma .err { color: #ebdbb2; background-color: #282828 } /* Error */
+.chroma .esc { color: #ebdbb2; background-color: #282828 } /* Escape */
+.chroma .g { color: #ebdbb2; background-color: #282828 } /* Generic */
+.chroma .k { color: #fe8019; background-color: #282828 } /* Keyword */
+.chroma .l { color: #ebdbb2; background-color: #282828 } /* Literal */
+.chroma .n { color: #ebdbb2; background-color: #282828 } /* Name */
+.chroma .o { color: #fe8019; background-color: #282828 } /* Operator */
+.chroma .x { color: #ebdbb2; background-color: #282828 } /* Other */
+.chroma .p { color: #ebdbb2; background-color: #282828 } /* Punctuation */
+.chroma .ch { color: #928374; font-style: italic; background-color: #282828 } /* Comment.Hashbang */
+.chroma .cm { color: #928374; font-style: italic; background-color: #282828 } /* Comment.Multiline */
+.chroma .cp { color: #8ec07c; background-color: #282828 } /* Comment.Preproc */
+.chroma .c1 { color: #928374; font-style: italic; background-color: #282828 } /* Comment.Single */
+.chroma .cs { color: #928374; font-style: italic; background-color: #282828 } /* Comment.Special */
+.chroma .gd { color: #282828; background-color: #fb4934 } /* Generic.Deleted */
+.chroma .ge { color: #83a598; text-decoration: underline; background-color: #282828 } /* Generic.Emph */
+.chroma .gr { color: #ebdbb2; font-weight: bold; background-color: #fb4934 } /* Generic.Error */
+.chroma .gh { color: #b8bb26; font-weight: bold; background-color: #282828 } /* Generic.Heading */
+.chroma .gi { color: #282828; background-color: #b8bb26 } /* Generic.Inserted */
+.chroma .go { color: #504945; background-color: #282828 } /* Generic.Output */
+.chroma .gp { color: #ebdbb2; background-color: #282828 } /* Generic.Prompt */
+.chroma .gs { color: #ebdbb2; background-color: #282828 } /* Generic.Strong */
+.chroma .gu { color: #b8bb26; font-weight: bold; background-color: #282828 } /* Generic.Subheading */
+.chroma .gt { color: #ebdbb2; font-weight: bold; background-color: #fb4934 } /* Generic.Traceback */
+.chroma .kc { color: #fe8019; background-color: #282828 } /* Keyword.Constant */
+.chroma .kd { color: #fe8019; background-color: #282828 } /* Keyword.Declaration */
+.chroma .kn { color: #fe8019; background-color: #282828 } /* Keyword.Namespace */
+.chroma .kp { color: #fe8019; background-color: #282828 } /* Keyword.Pseudo */
+.chroma .kr { color: #fe8019; background-color: #282828 } /* Keyword.Reserved */
+.chroma .kt { color: #fabd2f; background-color: #282828 } /* Keyword.Type */
+.chroma .ld { color: #ebdbb2; background-color: #282828 } /* Literal.Date */
+.chroma .m { color: #d3869b; background-color: #282828 } /* Literal.Number */
+.chroma .s { color: #b8bb26; background-color: #282828 } /* Literal.String */
+.chroma .na { color: #b8bb26; font-weight: bold; background-color: #282828 } /* Name.Attribute */
+.chroma .nb { color: #fabd2f; background-color: #282828 } /* Name.Builtin */
+.chroma .nc { color: #ebdbb2; background-color: #282828 } /* Name.Class */
+.chroma .no { color: #d3869b; background-color: #282828 } /* Name.Constant */
+.chroma .nd { color: #ebdbb2; background-color: #282828 } /* Name.Decorator */
+.chroma .ni { color: #fabd2f; background-color: #282828 } /* Name.Entity */
+.chroma .ne { color: #fb4934; background-color: #282828 } /* Name.Exception */
+.chroma .nf { color: #fabd2f; background-color: #282828 } /* Name.Function */
+.chroma .nl { color: #fb4934; background-color: #282828 } /* Name.Label */
+.chroma .nn { color: #ebdbb2; background-color: #282828 } /* Name.Namespace */
+.chroma .nx { color: #ebdbb2; background-color: #282828 } /* Name.Other */
+.chroma .py { color: #ebdbb2; background-color: #282828 } /* Name.Property */
+.chroma .nt { color: #fb4934; background-color: #282828 } /* Name.Tag */
+.chroma .nv { color: #ebdbb2; background-color: #282828 } /* Name.Variable */
+.chroma .ow { color: #fe8019; background-color: #282828 } /* Operator.Word */
+.chroma .w { color: #ebdbb2; background-color: #282828 } /* Text.Whitespace */
+.chroma .mb { color: #d3869b; background-color: #282828 } /* Literal.Number.Bin */
+.chroma .mf { color: #d3869b; background-color: #282828 } /* Literal.Number.Float */
+.chroma .mh { color: #d3869b; background-color: #282828 } /* Literal.Number.Hex */
+.chroma .mi { color: #d3869b; background-color: #282828 } /* Literal.Number.Integer */
+.chroma .mo { color: #d3869b; background-color: #282828 } /* Literal.Number.Oct */
+.chroma .sb { color: #b8bb26; background-color: #282828 } /* Literal.String.Backtick */
+.chroma .sc { color: #b8bb26; background-color: #282828 } /* Literal.String.Char */
+.chroma .sd { color: #b8bb26; background-color: #282828 } /* Literal.String.Doc */
+.chroma .s2 { color: #b8bb26; background-color: #282828 } /* Literal.String.Double */
+.chroma .se { color: #b8bb26; background-color: #282828 } /* Literal.String.Escape */
+.chroma .sh { color: #b8bb26; background-color: #282828 } /* Literal.String.Heredoc */
+.chroma .si { color: #b8bb26; background-color: #282828 } /* Literal.String.Interpol */
+.chroma .sx { color: #b8bb26; background-color: #282828 } /* Literal.String.Other */
+.chroma .sr { color: #b8bb26; background-color: #282828 } /* Literal.String.Regex */
+.chroma .s1 { color: #b8bb26; background-color: #282828 } /* Literal.String.Single */
+.chroma .ss { color: #83a598; background-color: #282828 } /* Literal.String.Symbol */
+.chroma .bp { color: #fabd2f; background-color: #282828 } /* Name.Builtin.Pseudo */
+.chroma .vc { color: #ebdbb2; background-color: #282828 } /* Name.Variable.Class */
+.chroma .vg { color: #ebdbb2; background-color: #282828 } /* Name.Variable.Global */
+.chroma .vi { color: #ebdbb2; background-color: #282828 } /* Name.Variable.Instance */
+.chroma .il { color: #d3869b; background-color: #282828 } /* Literal.Number.Integer.Long */
