+++ src/assets/css/_tachyons.css
@@ -1,0 +1,94 @@
+/*! TACHYONS v4.11.2 | http://tachyons.io */
+
+/*
+ *
+ *      ________            ______
+ *      ___  __/_____ _________  /______  ______________________
+ *      __  /  _  __ `/  ___/_  __ \_  / / /  __ \_  __ \_  ___/
+ *      _  /   / /_/ // /__ _  / / /  /_/ // /_/ /  / / /(__  )
+ *      /_/    \__,_/ \___/ /_/ /_/_\__, / \____//_/ /_//____/
+ *                                 /____/
+ *
+ *    TABLE OF CONTENTS
+ *
+ *    1. External Library Includes
+ *       - Normalize.css | http://normalize.css.github.io
+ *    2. Tachyons Modules
+ *    3. Variables
+ *       - Media Queries
+ *       - Colors
+ *    4. Debugging
+ *       - Debug all
+ *       - Debug children
+ *
+ */
+
+
+/* External Library Includes */
+@import 'tachyons/src/_normalize';
+
+
+/* Modules */
+@import 'tachyons/src/_box-sizing';
+@import 'tachyons/src/_aspect-ratios';
+@import 'tachyons/src/_images';
+@import 'tachyons/src/_background-size';
+@import 'tachyons/src/_background-position';
+@import 'tachyons/src/_outlines';
+@import 'tachyons/src/_borders';
+@import 'tachyons/src/_border-colors';
+@import 'tachyons/src/_border-radius';
+@import 'tachyons/src/_border-style';
+@import 'tachyons/src/_border-widths';
+@import 'tachyons/src/_box-shadow';
+@import 'tachyons/src/_code';
+@import 'tachyons/src/_coordinates';
+@import 'tachyons/src/_clears';
+@import 'tachyons/src/_display';
+@import 'tachyons/src/_flexbox';
+@import 'tachyons/src/_floats';
+@import 'tachyons/src/_font-family';
+@import 'tachyons/src/_font-style';
+@import 'tachyons/src/_font-weight';
+@import 'tachyons/src/_forms';
+@import 'tachyons/src/_heights';
+@import 'tachyons/src/_letter-spacing';
+@import 'tachyons/src/_line-height';
+@import 'tachyons/src/_links';
+@import 'tachyons/src/_lists';
+@import 'tachyons/src/_max-widths';
+@import 'tachyons/src/_widths';
+@import 'tachyons/src/_overflow';
+@import 'tachyons/src/_position';
+@import 'tachyons/src/_opacity';
+@import 'tachyons/src/_rotations';
+@import 'tachyons/src/_skins';
+@import 'tachyons/src/_skins-pseudo';
+@import 'tachyons/src/_spacing';
+@import 'tachyons/src/_negative-margins';
+@import 'tachyons/src/_tables';
+@import 'tachyons/src/_text-decoration';
+@import 'tachyons/src/_text-align';
+@import 'tachyons/src/_text-transform';
+@import 'tachyons/src/_type-scale';
+@import 'tachyons/src/_typography';
+@import 'tachyons/src/_utilities';
+@import 'tachyons/src/_visibility';
+@import 'tachyons/src/_white-space';
+@import 'tachyons/src/_vertical-align';
+@import 'tachyons/src/_hovers';
+@import 'tachyons/src/_z-index';
+@import 'tachyons/src/_nested';
+@import 'tachyons/src/_styles';
+
+/* Variables */
+/* Importing here will allow you to override any variables in the modules */
+@import 'tachyons/src/_colors';
+@import 'tachyons/src/_media-queries';
+
+/* Debugging */
+@import 'tachyons/src/_debug-children';
+@import 'tachyons/src/_debug-grid';
+
+/* Uncomment out the line below to help debug layout issues */
+/* @import 'tachyons/src/_debug'; */
