+++ src/assets/css/main.css
@@ -1,0 +1,21 @@
+@import '@resonate/tachyons';
+@import 'assets/css/_variables';
+@import 'assets/css/_typography';
+@import 'assets/css/_fills';
+@import 'assets/css/_code';
+@import 'assets/css/_hugo-internal-templates';
+@import 'assets/css/_social-icons';
+@import 'assets/css/_styles';
+@import 'assets/css/_syntax';
+
+@import 'assets/css/_forms';
+@import 'assets/css/_buttons';
+@import 'assets/css/_grid';
+@import 'assets/css/_header';
+@import 'assets/css/_footer';
+@import 'assets/css/_nav';
+
+@import '@resonate/search-component/index.css';
+
+@import 'assets/css/_marketing';
+@import 'assets/css/_manifesto';
