+++ src/layouts/_default/baseof.html
@@ -8,6 +8,5 @@
     {{ partial "components/footer" . }}
   </div>
   {{ partial "icons/icons" . }}
-  <script type="text/javascript" src="/dist/js/{{ index .Site.Data.js.hash "main.js" }}"></script>
 </body>
 </html>
