+++ src/layouts/partials/head.html
@@ -16,11 +16,27 @@
   <link rel="preload" href="https://static.resonate.is/fonts/Graphik-Regular.woff2" as="font" type="font/woff2" crossorigin>
   <link rel="preload" href="https://static.resonate.is/fonts/Graphik-Semibold.woff2" as="font" type="font/woff2" crossorigin>
 
-  <link rel="stylesheet" href="/dist/css/{{ index .Site.Data.css.hash "main.css" }}" />
+  {{ $css := resources.Get "css/main.css" | resources.PostCSS }}
+  {{ if hugo.IsProduction }}
+  {{ $css := $css | minify | fingerprint "sha512" | resources.PostProcess }}
+  <link rel="stylesheet" href="{{ $css.Permalink }}" integrity="{{ $css.Data.Integrity }}" />
+  {{ else }}
+  <link rel="stylesheet" href="{{ $css.Permalink }}" />
+  {{ end }}
 
   {{ if .IsTranslated -}}
     {{ range .Translations -}}
     <link rel="alternate" hreflang="{{ .Language.Lang }}" href="{{ .Permalink }}" title="{{ .Language.LanguageName }}">
     {{- end }}
+  {{ end }}
+
+  {{ $defines := dict "process.env.NODE_ENV" `"development"` "process.env.APP_HOST" `"beta.stream.resonate.coop"` }}
+  {{ $opts := dict "targetPath" "main.js" "defines" $defines }}
+  {{ $js := resources.Get "js/main.js" | js.Build $opts }}
+  {{ if hugo.IsProduction }}
+  {{ $js := $js | babel | minify | fingerprint "sha512" | resources.PostProcess }}
+  <script type="text/javascript" src="{{ $js.Permalink }}" integrity="{{ $js.Data.Integrity }}" defer></script>
+  {{else}}
+  <script type="text/javascript" src="{{ $js.Permalink }}" defer></script>
   {{ end }}
 </head>
