+++ src/package.json
@@ -17,11 +17,12 @@
     "nanoanimation": "^2.1.0",
     "nanochoo": "git+https://github.com/resonatecoop/nanochoo.git",
     "nanohtml": "^1.9.1",
-    "on-intersect": "^2.0.0"
+    "on-intersect": "^2.0.0",
+    "regenerator-runtime": "^0.13.9"
   },
   "devDependencies": {
     "@babel/core": "^7.2.2",
-    "@babel/preset-env": "^7.3.1",
+    "@babel/preset-env": "^7.15.6",
     "babelify": "^10.0.0",
     "browserify": "^16.2.3",
     "critical": "^2.0.0-27",
