+++ src/postcss.config.js
@@ -1,0 +1,19 @@
+module.exports = {
+  plugins: [
+    require('postcss-import')(),
+    require('postcss-preset-env')({
+      stage: 1,
+      features: {
+        browsers: ['last 1 version'],
+        'nesting-rules': true
+      }
+    }),
+    require('cssnano')({
+      preset: ['default', {
+        discardComments: {
+          removeAll: true
+        }
+      }]
+    })
+  ]
+}
